requirements.txt
streamlit
pillow
opencv-python
pytesseract
numpy
pandas
scikit-learn
matplotlib
sqlalchemy
python-dotenv
nltk
requests
